+++ x/common/testutil/mock/perp_market.go
@@ -3,10 +3,11 @@ package mock
 import (
 	time "time"
 
+	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	asset "github.com/NibiruChain/nibiru/x/common/asset"
 	"github.com/NibiruChain/nibiru/x/common/denoms"
 	v2types "github.com/NibiruChain/nibiru/x/perp/types/v2"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
 // default market with sensible values for tests